backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
opencv-python-headless
pillow
numpy
# Use pre-compiled dlib to avoid build issues
dlib-binary
face-recognition 
onnxruntime 
mtcnn
tensorflow
mediapipe